refactor: mvvm refactor + test mocks
