+++ apps/ledger-live-mobile/src/mvvm/features/LiveAppModal/__tests__/LiveAppModal.test.tsx
